Moved from Integer division to Double division for shop prices
Took 2 minutes

+++ generic/src/main/java/net/swofty/types/generic/shop/type/CoinShopPrice.java
@@ -45,7 +45,7 @@ public class CoinShopPrice implements ShopPrice {
 
     @Override
     public ShopPrice divide(double amount) {
-        return new CoinShopPrice(Math.max((int) (this.amount / amount), 1));
+        return new CoinShopPrice(Math.max((this.amount / amount), 1));
     }
 
 }
